requirement.txt
backcall==0.1.0
click==6.7
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
Flask-Script==2.0.6
Flask-WTF==0.14.2
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
kiwisolver==1.0.1
livereload==2.5.2
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.3.18
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
simplegeneric==0.8.1
six==1.11.0
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
